refactoring and fixing API
