+++ src/cryptoadvance/specter/api/rest/api.py
@@ -2,125 +2,63 @@ import logging
 import json
 
 import cryptoadvance.specter as specter
-from cryptoadvance.specter.api.rest.base import (
-    AdminResource,
-    BaseResource,
+from ...wallet import Wallet
+from .base import (
     SecureResource,
     rest_resource,
 )
-from cryptoadvance.specter.api.security import require_admin, verify_password
-from flask_restful import reqparse
+from ..security import require_admin, verify_password
+from flask_restful import abort
 from flask import current_app as app
 from datetime import datetime
 from ...specter_error import SpecterError
 
 from ...util.fee_estimation import get_fees
 
+from .. import auth
+from .resource_healthz import ResourceLiveness, ResourceReadyness
 from .resource_psbt import ResourcePsbt
+from .resource_specter import ResourceSpecter
+from .resource_txlist import ResourceTXlist
 
 logger = logging.getLogger(__name__)
-
-
-@rest_resource
-class ResourceSpecter(SecureResource):
-    """/api/v1alpha/specter"""
-
-    endpoints = ["/v1alpha/specter"]
-
-    def get(self, wallet_alias):
-        specter_data = app.specter
-
-        return_dict = {
-            "data_folder": specter_data.data_folder,
-            "config": specter_data.config,
-            "info": specter_data.info,
-            "network_info": specter_data.network_info,
-            "device_manager_datafolder": specter_data.device_manager.data_folder,
-            "devices_names": specter_data.device_manager.devices_names,
-            "wallets_names": specter_data.wallet_manager.wallets_names,
-            "last_update": datetime.now().strftime("%m/%d/%Y, %H:%M:%S"),
-        }
-
-        # Include alias list for easy lookup between names and alias
-        # Maybe there's a better way to create this
-        wallets_alias = []
-        alias_name = {}
-        name_alias = {}
-        for wallet in return_dict["wallets_names"]:
-            alias = specter_data.wallet_manager.wallets[wallet].alias
-            wallets_alias.append(alias)
-            alias_name[alias] = wallet
-            name_alias[wallet] = alias
-        return_dict["alias_name"] = alias_name
-        return_dict["name_alias"] = name_alias
-        return_dict["wallets_alias"] = wallets_alias
-        return return_dict
-
-
-@rest_resource
-class ResourceTXlist(SecureResource):
-    """/api/v1alpha/full_txlist"""
-
-    endpoints = ["/v1alpha/full_txlist/"]
-
-    def get(self):
-        try:
-            validate_merkle_proofs = app.specter.config.get("validate_merkle_proofs")
-            idx = 0
-            tx_len = 1
-            tx_list = []
-            while tx_len > 0:
-                transactions = app.specter.wallet_manager.full_txlist(
-                    idx, validate_merkle_proofs
-                )
-                tx_list.append(transactions)
-                tx_len = len(transactions)
-                idx += 1
-            # Flatten the list
-            flat_list = []
-            for element in tx_list:
-                for dic_item in element:
-                    flat_list.append(dic_item)
-
-        except SpecterError as se:
-            message = "API error: %s" % se
-            app.logger.error(message)
-            flat_list = message
-        return flat_list
 
 
 @rest_resource
 class ResourceWallet(SecureResource):
     """/api/v1alpha/wallet_info/<wallet_alias>/"""
 
-    endpoints = ["/v1alpha/wallet_info/<wallet_alias>/"]
+    endpoints = ["/v1alpha/wallets/<wallet_alias>/"]
 
     def get(self, wallet_alias):
+        user = auth.current_user()
         try:
-            wallet = app.specter.wallet_manager.get_by_alias(wallet_alias)
+            wallet: Wallet = app.specter.user_manager.get_user(
+                user
+            ).wallet_manager.get_by_alias(wallet_alias)
         except SpecterError as se:
-            message = "API error: %s" % se
-            app.logger.error(message)
-            return json.dumps(message)
+            logger.error(se)
+            if str(se).startswith(f"Wallet {wallet_alias} does not exist!"):
+                return abort(403, message=f"Wallet {wallet_alias} does not exist")
+            logger.error(se)
+            return abort(500)
 
         wallet.get_balance()
         wallet.check_utxo()
         wallet.check_unused()
+        logger.info("Muuh")
 
         return_dict = {}
         address_index = wallet.address_index
         validate_merkle_proofs = app.specter.config.get("validate_merkle_proofs")
-
+        logger.info("Meeh")
         tx_list = []
         idx = 0
         tx_len = 1
-        while tx_len > 0:
-            transactions = wallet.txlist(
-                idx, validate_merkle_proofs=validate_merkle_proofs
-            )
-            tx_list.append(transactions)
-            tx_len = len(transactions)
-            idx += 1
+        transactions = app.specter.wallet_manager.full_txlist(
+            fetch_transactions=False, validate_merkle_proofs=validate_merkle_proofs
+        )
+        tx_list.append(transactions)
 
         # Flatten the list
         flat_list = []
@@ -130,7 +68,7 @@ class ResourceWallet(SecureResource):
 
         # Check if scanning
         scan = wallet.rescan_progress
-
+        logger.info("Meeh3")
         return_dict[wallet_alias] = wallet.__dict__
         return_dict["txlist"] = flat_list
         return_dict["scan"] = scan
@@ -145,4 +83,4 @@ class ResourceWallet(SecureResource):
 
             return json.dumps(obj, default=default)
 
-        return safe_serialize(return_dict)
+        return json.loads(safe_serialize(return_dict))
