+++ src/cryptoadvance/specter/api/rest/resource_healthz.py
