+++ src/cryptoadvance/specter/api/templates/api.html
@@ -1,0 +1,18 @@
+<html>
+
+<body>
+The documentation for the API FOR THE MASTER BRANCH can be found here:
+<a href="https://github.com/cryptoadvance/specter-desktop/blob/master/docs/api.md">https://github.com/cryptoadvance/specter-desktop/blob/master/docs/api.md</a>
+
+<br><br>
+Make sure to choose the right version for the version you're currently using: {{version}}
+
+<br><br>
+{% if version != "custom" %}
+So assuming that, please use this url:
+<a href="https://github.com/cryptoadvance/specter-desktop/blob/v{{version}}/docs/api.md">https://github.com/cryptoadvance/specter-desktop/blob/v{{version}}/docs/api.md</a>
+{% endif %}
+
+</body>
+
+</html>
