+++ src/cryptoadvance/specter/api/views.py
@@ -6,9 +6,12 @@ Use this to display server debugging info during development
 """
 from flask import render_template
 from cryptoadvance.specter.api import api_bp
+from flask import current_app as app
 
 
 @api_bp.route("/")
 def api():
     """rendering the documentation for the api"""
-    return render_template("api.html", msg="API Blueprint View")
+    return render_template(
+        "api.html", version=app.specter.version.get_current_version()
+    )
