+++ src/cryptoadvance/specter/config.py
@@ -38,6 +38,7 @@ class BaseConfig(object):
     SPECTER_DATA_FOLDER = os.path.expanduser(
         os.getenv("SPECTER_DATA_FOLDER", "~/.specter")
     )
+    SPECTER_API_ACTIVE = _get_bool_env_var("SPECTER_API_ACTIVE", "False")
     # Logging
     # SPECTER_LOGFILE will get created dynamically in server.py
     # using:
@@ -111,10 +112,14 @@ class DevelopmentConfig(BaseConfig):
     SPECTER_DATA_FOLDER = os.path.expanduser(
         os.getenv("SPECTER_DATA_FOLDER", "~/.specter_dev")
     )
+    # API active by default in dev-mode
+    SPECTER_API_ACTIVE = _get_bool_env_var("SPECTER_API_ACTIVE", "True")
 
 
 class TestConfig(BaseConfig):
     SECRET_KEY = "test key"
+    # API active by default in test-mode
+    SPECTER_API_ACTIVE = _get_bool_env_var("SPECTER_API_ACTIVE", "True")
 
 
 class CypressTestConfig(TestConfig):
