+++ tests/test_rest.py
@@ -1,3 +1,4 @@
+import pytest
 import json
 import base64
 import logging
